fix(tests): isolate navItems to avoid Deno-Windows promise leak
Extract the static navItems catalogue into js/src/web/nav-items.js
and re-export it from views.js. The two new tests now import from
the lightweight module so they no longer drag React, dom.js, and
the rest of the SPA module subtree into Deno's strict test runner,
which was leaking a microtask on Windows ("Promise resolution is
still pending but the event loop has already resolved").

Verified locally: deno test passes 200/200 with no leak.
