+++ js/tests/connection-guides.test.js
@@ -9,7 +9,7 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
 
-import { navItems } from '../src/web/views.js';
+import { navItems } from '../src/web/nav-items.js';
 import {
   connectionGuides,
   providerCatalogue,
